@@ -3,8 +3,8 @@ name = "medea-turn"
 version = "0.7.0-dev"
 authors = ["Instrumentisto Team <developer@instrumentisto.com>"]
 edition = "2021"
-rust-version = "1.70"
-description = "TURN implementation used by Medea media server."
+rust-version = "1.74"
+description = "STUN/TURN server implementation used by Medea media server."
 license = "MIT OR Apache-2.0"
 homepage = "https://github.com/instrumentisto/medea-turn-rs"
 repository = "https://github.com/instrumentisto/medea-turn-rs"
@@ -14,11 +14,11 @@ publish = false
 async-trait = "0.1"
 bytecodec = "0.4.15"
 bytes = "1.6"
+derive_more = { version = "1.0.0-beta.6", features = ["debug", "display", "error", "from"] }
 futures = "0.3"
 log = "0.4"
 rand = "0.8"
 stun_codec = "0.3"
-thiserror = "1.0"
 tokio = { version = "1.32", default-features = false, features = ["io-util", "macros", "net", "rt-multi-thread", "time"] }
 tokio-util = { version = "0.7", features = ["codec"] }
 
