@@ -1,7 +1,4 @@
-//! [`RelayAllocator`] is used to create relay transports wit the given
-//! configuration.
-
-#![allow(clippy::module_name_repetitions)]
+//! Relay definitions.
 
 use std::{
     net::{IpAddr, SocketAddr},
@@ -9,39 +6,39 @@ use std::{
 };
 use tokio::net::UdpSocket;
 
-use crate::{con, Error};
+use crate::{transport, Error};
 
-/// [`RelayAllocator`] is used to generate a Relay Address when creating an
-/// allocation.
-#[derive(Debug)]
-pub struct RelayAllocator {
-    /// `relay_address` is the IP returned to the user when the relay is
-    /// created.
+/// Generator of relay addresses when creating an [allocation].
+///
+/// [allocation]: https://tools.ietf.org/html/rfc5766#section-5
+#[derive(Clone, Debug)]
+pub struct Allocator {
+    /// [`IpAddr`] returned to the user when a relay is created.
     pub relay_address: IpAddr,
 
-    /// `min_port` the minimum port to allocate.
+    /// Minimum (inclusive) port to allocate.
     pub min_port: u16,
 
-    /// `max_port` the maximum (inclusive) port to allocate.
+    /// Maximum (inclusive) port to allocate.
     pub max_port: u16,
 
-    /// `max_retries` the amount of tries to allocate a random port in the
-    /// defined range.
+    /// Amount of tries to allocate a random port in the allowed range.
     pub max_retries: u16,
 
-    /// `address` is passed to Listen/ListenPacket when creating the Relay.
+    /// Address passed when creating a relay.
     pub address: String,
 }
 
-impl RelayAllocator {
+impl Allocator {
     /// Allocates a new relay connection.
     ///
     /// # Errors
     ///
-    /// With [`Error::MaxRetriesExceeded`] if the requested port is `0` and
-    /// failed to find a free port in the specified maximum retries.
+    /// - With an [`Error::MaxRetriesExceeded`] if the requested port is `0` and
+    /// failed to find a free port in the specified [`max_retries`].
+    /// - With an [`Error::Transport`] if failed to bind to the specified port.
     ///
-    /// With [`Error::Io`] if failed to bind to the specified port.
+    /// [`max_retries`]: Allocator::max_retries
     pub async fn allocate_conn(
         &self,
         use_ipv4: bool,
@@ -55,29 +52,33 @@ impl RelayAllocator {
                 let port = self.min_port
                     + rand::random::<u16>()
                         % (self.max_port - self.min_port + 1);
-                let addr = con::lookup_host(
+                let addr = transport::lookup_host(
                     use_ipv4,
-                    &format!("{}:{}", self.address, port),
+                    &format!("{}:{port}", self.address),
                 )
                 .await?;
                 let Ok(conn) = UdpSocket::bind(addr).await else {
                     continue;
                 };
 
-                let mut relay_addr = conn.local_addr()?;
+                let mut relay_addr =
+                    conn.local_addr().map_err(transport::Error::from)?;
                 relay_addr.set_ip(self.relay_address);
                 return Ok((Arc::new(conn), relay_addr));
             }
 
             Err(Error::MaxRetriesExceeded)
         } else {
-            let addr = con::lookup_host(
+            let addr = transport::lookup_host(
                 use_ipv4,
-                &format!("{}:{}", self.address, requested_port),
+                &format!("{}:{requested_port}", self.address),
             )
             .await?;
-            let conn = Arc::new(UdpSocket::bind(addr).await?);
-            let mut relay_addr = conn.local_addr()?;
+            let conn = Arc::new(
+                UdpSocket::bind(addr).await.map_err(transport::Error::from)?,
+            );
+            let mut relay_addr =
+                conn.local_addr().map_err(transport::Error::from)?;
             relay_addr.set_ip(self.relay_address);
 
             Ok((conn, relay_addr))
