@@ -1,47 +1,42 @@
-//! TURN [`Channel`].
+//! [Channel] definitions.
 //!
-//! [`Channel`]: https://tools.ietf.org/html/rfc5766#section-2.5
+//! [Channel]: https://tools.ietf.org/html/rfc5766#section-2.5
 
-use std::{collections::HashMap, net::SocketAddr, sync::Arc};
+use std::{collections::HashMap, net::SocketAddr, sync::Arc, time::Duration};
 
 use tokio::{
     sync::{mpsc, Mutex},
-    time::{sleep, Duration, Instant},
+    time::{sleep, Instant},
 };
 
-/// TURN [`Channel`].
+/// Representation of a [channel].
 ///
-/// [`Channel`]: https://tools.ietf.org/html/rfc5766#section-2.5
-#[derive(Clone)]
+/// [channel]: https://tools.ietf.org/html/rfc5766#section-2.5
+#[derive(Clone, Debug)]
 pub(crate) struct ChannelBind {
-    /// Transport address of the peer.
+    /// Transport address of the peer behind this [`ChannelBind`].
     peer: SocketAddr,
 
-    /// Channel number.
+    /// Number of this [`ChannelBind`].
     number: u16,
 
-    /// Channel to the internal loop used to update lifetime or drop channel
-    /// binding.
-    reset_tx: Option<mpsc::Sender<Duration>>,
+    /// [`mpsc::Sender`] to the internal loop of this [`ChannelBind`], used to
+    /// update its lifetime or stop it.
+    reset_tx: mpsc::Sender<Duration>,
 }
 
 impl ChannelBind {
-    /// Creates a new [`ChannelBind`]
-    pub(crate) const fn new(number: u16, peer: SocketAddr) -> Self {
-        Self { number, peer, reset_tx: None }
-    }
-
-    /// Starts [`ChannelBind`]'s internal lifetime watching loop.
-    pub(crate) fn start(
-        &mut self,
+    /// Creates a new [`ChannelBind`] and [`spawn`]s a loop watching its
+    /// lifetime.
+    ///
+    /// [`spawn`]: tokio::spawn()
+    pub(crate) fn new(
+        number: u16,
+        peer: SocketAddr,
         bindings: Arc<Mutex<HashMap<u16, Self>>>,
         lifetime: Duration,
-    ) {
+    ) -> Self {
         let (reset_tx, mut reset_rx) = mpsc::channel(1);
-        self.reset_tx = Some(reset_tx);
-
-        let number = self.number;
-
         drop(tokio::spawn(async move {
             let timer = sleep(lifetime);
             tokio::pin!(timer);
@@ -51,7 +46,8 @@ impl ChannelBind {
                     () = &mut timer => {
                         if bindings.lock().await.remove(&number).is_none() {
                             log::error!(
-                                "Failed to remove ChannelBind for {number}"
+                                "Failed to remove \
+                                 `ChannelBind(number: {number})`",
                             );
                         }
                         break;
@@ -66,44 +62,50 @@ impl ChannelBind {
                 }
             }
         }));
-    }
 
-    /// Returns transport address of the peer.
+        Self { peer, number, reset_tx }
+    }
+    /// Returns the [`SocketAddr`] of the peer behind this [`ChannelBind`].
     pub(crate) const fn peer(&self) -> SocketAddr {
         self.peer
     }
 
-    /// Returns channel number.
+    /// Returns the number of this [`ChannelBind`].
     pub(crate) const fn num(&self) -> u16 {
         self.number
     }
 
-    /// Updates [`ChannelBind`]'s lifetime.
+    /// Updates the `lifetime` of this [`ChannelBind`].
     pub(crate) async fn refresh(&self, lifetime: Duration) {
-        if let Some(tx) = &self.reset_tx {
-            _ = tx.send(lifetime).await;
-        }
+        _ = self.reset_tx.send(lifetime).await;
     }
 }
 
 #[cfg(test)]
-mod channel_bind_test {
-    use std::net::Ipv4Addr;
+mod allocation_spec {
+    use std::{
+        net::{Ipv4Addr, SocketAddr},
+        sync::Arc,
+        time::Duration,
+    };
 
     use tokio::net::UdpSocket;
 
     use crate::{
-        allocation::Allocation,
         attr::{ChannelNumber, Username},
-        con, Error, FiveTuple,
+        server::DEFAULT_LIFETIME,
+        Allocation, Error, FiveTuple,
     };
 
-    use super::*;
+    #[cfg(doc)]
+    use super::ChannelBind;
 
-    async fn create_channel_bind(
+    /// Creates an [`Allocation`] with a bound [`ChannelBind`] for testing
+    /// purposes.
+    async fn create_channel_bind_allocation(
         lifetime: Duration,
     ) -> Result<Allocation, Error> {
-        let turn_socket = Arc::new(UdpSocket::bind("0.0.0.0:0").await?);
+        let turn_socket = Arc::new(UdpSocket::bind("0.0.0.0:0").await.unwrap());
         let relay_socket = Arc::clone(&turn_socket);
         let relay_addr = relay_socket.local_addr().unwrap();
         let a = Allocation::new(
@@ -111,6 +113,7 @@ mod channel_bind_test {
             relay_socket,
             relay_addr,
             FiveTuple::default(),
+            DEFAULT_LIFETIME,
             Username::new(String::from("user")).unwrap(),
             None,
         );
@@ -123,12 +126,14 @@ mod channel_bind_test {
     }
 
     #[tokio::test]
-    async fn test_channel_bind() {
-        let a = create_channel_bind(Duration::from_millis(20)).await.unwrap();
+    async fn channel_bind_is_present() {
+        let a = create_channel_bind_allocation(Duration::from_millis(20))
+            .await
+            .unwrap();
 
         let result = a.get_channel_addr(&ChannelNumber::MIN).await;
         if let Some(addr) = result {
-            assert_eq!(addr.ip().to_string(), "0.0.0.0");
+            assert_eq!(addr.ip().to_string(), "0.0.0.0", "wrong IP address");
         } else {
             panic!("expected some, but got none");
         }
