@@ -1,7 +1,7 @@
-//! TURN [Allocation] [Permission].
+//! [Allocation] [permission] definitions.
 //!
-//! [Allocation]: https://datatracker.ietf.org/doc/html/rfc5766#section-2.2
-//! [Permission]: https://datatracker.ietf.org/doc/html/rfc5766#section-8
+//! [Allocation]: https://tools.ietf.org/html/rfc5766#section-2.2
+//! [permission]: https://tools.ietf.org/html/rfc5766#section-8
 
 use std::{collections::HashMap, net::IpAddr, sync::Arc};
 
@@ -10,40 +10,38 @@ use tokio::{
     time::{sleep, Duration, Instant},
 };
 
-/// The Permission Lifetime MUST be 300 seconds (= 5 minutes)[1].
+/// [Lifetime][1] of a [`Permission`].
 ///
-/// [1]: https://datatracker.ietf.org/doc/html/rfc5766#section-8
+/// > The Permission Lifetime MUST be 300 seconds (= 5 minutes).
+///
+/// [1]: https://tools.ietf.org/html/rfc5766#section-8
 pub(crate) const PERMISSION_LIFETIME: Duration = Duration::from_secs(5 * 60);
 
-/// TURN [Allocation] [Permission].
+/// Representation of an [allocation] [permission].
 ///
-/// [Allocation]: https://datatracker.ietf.org/doc/html/rfc5766#section-2.2
-/// [Permission]: https://datatracker.ietf.org/doc/html/rfc5766#section-8
+/// [allocation]: https://tools.ietf.org/html/rfc5766#section-2.2
+/// [permission]: https://tools.ietf.org/html/rfc5766#section-8
+#[derive(Debug)]
 pub(crate) struct Permission {
-    /// [`IpAddr`] of this permission that is matched with the source IP
+    /// [`IpAddr`] of this [`Permission`] that is matched with the source IP
     /// address of the datagram received.
     ip: IpAddr,
 
-    /// Channel to the inner lifetime watching loop.
-    reset_tx: Option<mpsc::Sender<Duration>>,
+    /// [`mpsc::Sender`] to the inner lifetime watching loop.
+    reset_tx: mpsc::Sender<Duration>,
 }
 
 impl Permission {
-    /// Creates a new [`Permission`].
-    pub(crate) const fn new(ip: IpAddr) -> Self {
-        Self { ip, reset_tx: None }
-    }
-
-    /// Starts [`Permission`]'s internal lifetime watching loop.
-    pub(crate) fn start(
-        &mut self,
+    /// Creates a new [`Permission`] and [`spawn`]s a loop watching its
+    /// lifetime.
+    ///
+    /// [`spawn`]: tokio::spawn()
+    pub(crate) fn new(
+        ip: IpAddr,
         permissions: Arc<Mutex<HashMap<IpAddr, Self>>>,
         lifetime: Duration,
-    ) {
+    ) -> Self {
         let (reset_tx, mut reset_rx) = mpsc::channel(1);
-        self.reset_tx = Some(reset_tx);
-
-        let ip = self.ip;
 
         drop(tokio::spawn(async move {
             let timer = sleep(lifetime);
@@ -65,17 +63,17 @@ impl Permission {
                 }
             }
         }));
+
+        Self { ip, reset_tx }
     }
 
-    /// Returns [`IpAddr`] of this [`Permission`].
+    /// Returns the [`IpAddr`] of this [`Permission`].
     pub(crate) const fn ip(&self) -> IpAddr {
         self.ip
     }
 
-    /// Updates [`Permission`]'s lifetime.
+    /// Updates the `lifetime` of this [`Permission`].
     pub(crate) async fn refresh(&self, lifetime: Duration) {
-        if let Some(tx) = &self.reset_tx {
-            _ = tx.send(lifetime).await;
-        }
+        _ = self.reset_tx.send(lifetime).await;
     }
 }
