@@ -1,6 +1,10 @@
-//! [`ChannelData`] message implementation.
+//! [TURN ChannelData Message][1] implementation.
+//!
+//! [1]: https://tools.ietf.org/html/rfc5766#section-11.4
 
-use crate::{attr::ChannelNumber, Error};
+use derive_more::{Display, Error};
+
+use crate::attr::ChannelNumber;
 
 /// [`ChannelData`] message MUST be padded to a multiple of four bytes in order
 /// to ensure the alignment of subsequent messages.
@@ -8,74 +12,82 @@ const PADDING: usize = 4;
 
 /// [Channel Number] field size.
 ///
-/// [Channel Number]: https://datatracker.ietf.org/doc/html/rfc5766#section-11.4
-const CHANNEL_DATA_NUMBER_SIZE: usize = 2;
+/// [Channel Number]: https://tools.ietf.org/html/rfc5766#section-11.4
+const NUMBER_SIZE: usize = 2;
 
 /// [Length] field size.
 ///
-/// [Length]: https://datatracker.ietf.org/doc/html/rfc5766#section-11.4
-const CHANNEL_DATA_LENGTH_SIZE: usize = 2;
+/// [Length]: https://tools.ietf.org/html/rfc5766#section-11.4
+const LENGTH_SIZE: usize = 2;
 
-/// [ChannelData] message header size.
+/// [ChannelData Message][1] header size.
 ///
-/// [ChannelData]: https://datatracker.ietf.org/doc/html/rfc5766#section-11.4
-const CHANNEL_DATA_HEADER_SIZE: usize =
-    CHANNEL_DATA_LENGTH_SIZE + CHANNEL_DATA_NUMBER_SIZE;
+/// [1]: https://tools.ietf.org/html/rfc5766#section-11.4
+const HEADER_SIZE: usize = LENGTH_SIZE + NUMBER_SIZE;
 
-/// [`ChannelData`] represents the `ChannelData` Message defined in
-/// [RFC 5766](https://www.rfc-editor.org/rfc/rfc5766#section-11.4).
+/// Representation of [TURN ChannelData Message][1] defined in [RFC 5766].
+///
+/// [1]: https://tools.ietf.org/html/rfc5766#section-11.4
+/// [RFC 5766]: https://tools.ietf.org/html/rfc5766
 #[derive(Debug)]
-pub(crate) struct ChannelData {
-    /// Parsed [`ChannelData`] [Channel Number][1].
+pub struct ChannelData {
+    /// Parsed [Channel Number][1].
     ///
-    /// [1]: https://datatracker.ietf.org/doc/html/rfc5766#section-11.4
+    /// [1]: https://tools.ietf.org/html/rfc5766#section-11.4
     number: u16,
 
-    /// Parsed [`ChannelData`] payload.
+    /// Parsed payload.
     data: Vec<u8>,
 }
 
 impl ChannelData {
-    /// Returns `true` if `buf` looks like the `ChannelData` Message.
-    #[allow(clippy::missing_asserts_for_indexing)] // Length is checked
-    pub(crate) fn is_channel_data(buf: &[u8]) -> bool {
-        if buf.len() < CHANNEL_DATA_HEADER_SIZE {
+    /// Checks whether the provided `data` represents a [`ChannelData`] message.
+    pub(crate) fn is_channel_data(data: &[u8]) -> bool {
+        // PANIC: Indexing is OK here, since the length is checked with the
+        //        first `if` expression.
+        #![allow(clippy::missing_asserts_for_indexing)] // false positive
+
+        if data.len() < HEADER_SIZE {
             return false;
         }
         let len = usize::from(u16::from_be_bytes([
-            buf[CHANNEL_DATA_NUMBER_SIZE],
-            buf[CHANNEL_DATA_NUMBER_SIZE + 1],
+            data[NUMBER_SIZE],
+            data[NUMBER_SIZE + 1],
         ]));
 
-        if len > buf[CHANNEL_DATA_HEADER_SIZE..].len() {
+        if len > data[HEADER_SIZE..].len() {
             return false;
         }
 
-        ChannelNumber::new(u16::from_be_bytes([buf[0], buf[1]])).is_ok()
+        ChannelNumber::new(u16::from_be_bytes([data[0], data[1]])).is_ok()
     }
 
-    /// Decodes the given raw message as [`ChannelData`].
-    pub(crate) fn decode(mut raw: Vec<u8>) -> Result<Self, Error> {
-        if raw.len() < CHANNEL_DATA_HEADER_SIZE {
-            return Err(Error::UnexpectedEof);
+    /// Decodes the provided `raw` message as a [`ChannelData`] message.
+    ///
+    /// # Errors
+    ///
+    /// See the [`FormatError`] for details.
+    pub(crate) fn decode(mut raw: Vec<u8>) -> Result<Self, FormatError> {
+        if raw.len() < HEADER_SIZE {
+            return Err(FormatError::BadChannelDataLength);
         }
 
         let number = u16::from_be_bytes([raw[0], raw[1]]);
         if ChannelNumber::new(number).is_err() {
-            return Err(Error::InvalidChannelNumber);
+            return Err(FormatError::InvalidChannelNumber);
         }
 
         let l = usize::from(u16::from_be_bytes([
-            raw[CHANNEL_DATA_NUMBER_SIZE],
-            raw[CHANNEL_DATA_NUMBER_SIZE + 1],
+            raw[NUMBER_SIZE],
+            raw[NUMBER_SIZE + 1],
         ]));
 
-        if l > raw[CHANNEL_DATA_HEADER_SIZE..].len() {
-            return Err(Error::BadChannelDataLength);
+        if l > raw[HEADER_SIZE..].len() {
+            return Err(FormatError::BadChannelDataLength);
         }
 
         // Discard header and padding.
-        drop(raw.drain(0..CHANNEL_DATA_HEADER_SIZE));
+        drop(raw.drain(0..HEADER_SIZE));
         if l != raw.len() {
             raw.truncate(l);
         }
@@ -83,45 +95,44 @@ impl ChannelData {
         Ok(Self { data: raw, number })
     }
 
-    /// Returns [`ChannelData`] [Channel Number][1].
+    /// Returns payload of this [`ChannelData`] message.
+    pub(crate) fn data(self) -> Vec<u8> {
+        self.data
+    }
+
+    /// Returns [Channel Number][1] of this [`ChannelData`] message.
     ///
-    /// [1]: https://datatracker.ietf.org/doc/html/rfc5766#section-11.4
+    /// [1]: https://tools.ietf.org/html/rfc5766#section-11.4
     pub(crate) const fn num(&self) -> u16 {
         self.number
     }
 
-    /// Encodes the provided [`ChannelData`] payload and channel number to
-    /// bytes.
+    /// Encodes the provided `payload` and [Channel Number][1] as
+    /// [`ChannelData`] message bytes.
+    ///
+    /// [1]: https://tools.ietf.org/html/rfc5766#section-11.4
     pub(crate) fn encode(
-        mut data: Vec<u8>,
+        payload: &[u8],
         chan_num: u16,
-    ) -> Result<Vec<u8>, Error> {
-        #[allow(clippy::map_err_ignore)]
-        let len = u16::try_from(data.len())
-            .map_err(|_| Error::BadChannelDataLength)?;
-        for i in len.to_be_bytes().into_iter().rev() {
-            data.insert(0, i);
-        }
-        for i in chan_num.to_be_bytes().into_iter().rev() {
-            data.insert(0, i);
-        }
+    ) -> Result<Vec<u8>, FormatError> {
+        let length = HEADER_SIZE + payload.len();
+        let padded_length = nearest_padded_value_length(length);
 
-        let padded = nearest_padded_value_length(data.len());
-        let bytes_to_add = padded - data.len();
-        if bytes_to_add > 0 {
-            data.extend_from_slice(&vec![0; bytes_to_add]);
-        }
+        #[allow(clippy::map_err_ignore)] // intentional
+        let len = u16::try_from(payload.len())
+            .map_err(|_| FormatError::BadChannelDataLength)?;
 
-        Ok(data)
-    }
+        let mut encoded = vec![0u8; padded_length];
 
-    /// Returns [`ChannelData`] payload.
-    pub(crate) fn data(self) -> Vec<u8> {
-        self.data
+        encoded[..NUMBER_SIZE].copy_from_slice(&chan_num.to_be_bytes());
+        encoded[NUMBER_SIZE..HEADER_SIZE].copy_from_slice(&len.to_be_bytes());
+        encoded[HEADER_SIZE..length].copy_from_slice(payload);
+
+        Ok(encoded)
     }
 }
 
-/// Calculates nearest padded length for the [`ChannelData`].
+/// Calculates a nearest padded length for a [`ChannelData`] message.
 pub(crate) const fn nearest_padded_value_length(l: usize) -> usize {
     let mut n = PADDING * (l / PADDING);
     if n < l {
@@ -130,29 +141,43 @@ pub(crate) const fn nearest_padded_value_length(l: usize) -> usize {
     n
 }
 
+/// Possible errors of a [`ChannelData`] message format.
+#[derive(Clone, Copy, Debug, Display, Error, Eq, PartialEq)]
+pub enum FormatError {
+    /// [Channel Number][1] is incorrect.
+    ///
+    /// [1]: https://tools.ietf.org/html/rfc5766#section-11.4
+    #[display("Channel Number not in [0x4000, 0x7FFF]")]
+    InvalidChannelNumber,
+
+    /// Incorrect message length.
+    #[display("Invalid `ChannelData` length")]
+    BadChannelDataLength,
+}
+
 #[cfg(test)]
-mod chandata_test {
-    use super::*;
+mod spec {
+    use crate::attr::ChannelNumber;
+
+    use super::{ChannelData, FormatError};
 
     #[test]
-    fn test_channel_data_encode() {
+    fn encodes() {
         let encoded =
-            ChannelData::encode(vec![1, 2, 3, 4], ChannelNumber::MIN + 1)
-                .unwrap();
+            ChannelData::encode(&[1, 2, 3, 4], ChannelNumber::MIN + 1).unwrap();
         let decoded = ChannelData::decode(encoded.clone()).unwrap();
 
         assert!(
             ChannelData::is_channel_data(&encoded),
-            "unexpected IsChannelData"
+            "wrong `is_channel_data`",
         );
-
-        assert_eq!(vec![1, 2, 3, 4], decoded.data, "not equal");
-        assert_eq!(ChannelNumber::MIN + 1, decoded.number, "not equal");
+        assert_eq!(vec![1, 2, 3, 4], decoded.data, "wrong decoded data");
+        assert_eq!(ChannelNumber::MIN + 1, decoded.number, "wrong number");
     }
 
     #[test]
-    fn test_channel_data_equal() {
-        let tests = vec![
+    fn encoded_equality() {
+        let tests = [
             (
                 "equal",
                 ChannelData { number: ChannelNumber::MIN, data: vec![1, 2, 3] },
@@ -186,55 +211,52 @@ mod chandata_test {
         ];
 
         for (name, a, b, r) in tests {
-            let v = ChannelData::encode(a.data.clone(), a.number)
-                == ChannelData::encode(b.data.clone(), b.number);
-            assert_eq!(v, r, "unexpected: ({name}) {r} != {r}");
+            let v = ChannelData::encode(&a.data, a.number)
+                == ChannelData::encode(&b.data, b.number);
+
+            assert_eq!(v, r, "wrong equality of {name}");
         }
     }
 
     #[test]
-    fn test_channel_data_decode() {
-        let tests = vec![
-            ("small", vec![1, 2, 3], Error::UnexpectedEof),
+    fn fails_decoding_correctly() {
+        let tests = [
+            ("small", vec![1, 2, 3], FormatError::BadChannelDataLength),
             (
                 "zeroes",
                 vec![0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                Error::InvalidChannelNumber,
+                FormatError::InvalidChannelNumber,
             ),
             (
                 "bad chan number",
                 vec![63, 255, 0, 0, 0, 4, 0, 0, 1, 2, 3, 4],
-                Error::InvalidChannelNumber,
+                FormatError::InvalidChannelNumber,
             ),
             (
                 "bad length",
                 vec![0x40, 0x40, 0x02, 0x23, 0x16, 0, 0, 0, 0, 0, 0, 0],
-                Error::BadChannelDataLength,
+                FormatError::BadChannelDataLength,
             ),
         ];
-
         for (name, buf, want_err) in tests {
-            if let Err(err) = ChannelData::decode(buf) {
-                assert_eq!(
-                    want_err, err,
-                    "unexpected: ({name}) {want_err} != {err}"
-                );
+            if let Err(e) = ChannelData::decode(buf) {
+                assert_eq!(want_err, e, "wrong error of {name}");
             } else {
-                panic!("expected error, but got ok");
+                panic!("expected `Err`, but got `Ok` in {name}");
             }
         }
     }
 
     #[test]
-    fn test_is_channel_data() {
-        let tests = vec![
+    fn is_channel_data_detects_correctly() {
+        let tests = [
             ("small", vec![1, 2, 3, 4], false),
             ("zeroes", vec![0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0], false),
         ];
-
         for (name, buf, r) in tests {
             let v = ChannelData::is_channel_data(&buf);
-            assert_eq!(v, r, "unexpected: ({name}) {r} != {v}");
+
+            assert_eq!(v, r, "wrong result in {name}");
         }
     }
 
@@ -261,11 +283,11 @@ mod chandata_test {
     ];
 
     #[test]
-    fn test_chrome_channel_data() {
+    fn chrome_channel_data() {
         let mut data = vec![];
         let mut messages = vec![];
 
-        // Decoding hex data into binary.
+        // Decoding HEX data into binary.
         for h in &CHANDATA_TEST_HEX {
             let b = match hex::decode(h) {
                 Ok(b) => b,
@@ -274,21 +296,20 @@ mod chandata_test {
             data.push(b);
         }
 
-        // All hex streams decoded to raw binary format and stored in data
+        // All HEX streams decoded to raw binary format and stored in the `data`
         // slice. Decoding packets to messages.
         for packet in data {
             let m = ChannelData::decode(packet.clone()).unwrap();
 
-            let encoded =
-                ChannelData::encode(m.data.clone(), m.number).unwrap();
+            let encoded = ChannelData::encode(&m.data, m.number).unwrap();
             let decoded = ChannelData::decode(encoded.clone()).unwrap();
 
-            assert_eq!(m.data, decoded.data, "should be equal");
-            assert_eq!(m.number, decoded.number, "should be equal");
+            assert_eq!(m.data, decoded.data, "wrong payload");
+            assert_eq!(m.number, decoded.number, "wrong number");
 
             messages.push(m);
         }
 
-        assert_eq!(messages.len(), 2, "unexpected message slice list");
+        assert_eq!(messages.len(), 2, "wrong number of messages");
     }
 }
